requirements.txt
aiohttp                 ==3.8.4
aiosignal               ==1.3.1
async-timeout           ==4.0.2
attrs                   ==23.1.0
backoff                 ==2.2.1
branca                  ==0.6.0
certifi                 ==2023.7.22
chardet                 ==4.0.0
charset-normalizer      ==3.1.0
discord-py-interactions ==4.4.0
discord-typings         ==0.5.1
dnspython               ==2.3.0
emoji                   ==2.2.0
folium                  ==0.14.0
frozenlist              ==1.3.3
idna                    ==3.4
Jinja2                  ==3.1.2
MarkupSafe              ==2.1.3
multidict               ==6.0.4
numpy                   ==1.24.3
opencage                ==2.3.0
pandas                  ==2.0.1
phonenumbers            ==8.13.18
pip                     ==23.2.1
py-cord                 ==2.4.1
pymongo                 ==4.3.3
python-dateutil         ==2.8.2
python-dotenv           ==1.0.0
pytz                    ==2023.3
requests                ==2.31.0
six                     ==1.16.0
tomli                   ==2.0.1
typing_extensions       ==4.5.0
tzdata                  ==2023.3
urllib3                 ==2.0.4
yarl                    ==1.9.2